Refactoring frontend for good dev practices
